Set up test comparisons, fill out a bunch of instruction functions, other crap in between

+++ src/decoder.h
@@ -7,5 +7,5 @@
 #include <string>
 
 namespace Decoder {
-  Instruction decode(const uint8_t* mem, uint8_t addr);
+  Instruction decode(const uint8_t* mem, uint16_t addr);
 };
